Update patch set 1
I would prefer that you didn't submit this.

The change is probably good, but now is not the right time to make a
global search-and-replace for a class name. With a major code drop
coming soon and some active work going on in an environment that is
still using the previous names, this change would guarantee some severe
merge pain for months to come.

The right time to make this change will be when all development happens
directly in git, just before a stabilization branch is separated from
the trunk.

Patch-set: 1