Allow easy use of IME on devices with hardware keyboards
When a user has the hardware keyboard hidden on a device that supports
this, allow holding the Menu button to bring up the currently selected
IME.

This change gives users of devices with no keyboards and users with
devices that have hidden keyboards experience parity. Note that the
Search button currently has exactly the same behavior, but there
does not appear to be a way to press the search button when the
keyboard is closed on existing devices.