Update patch set 1
Patch Set 1: No score

Since line 651 checks for the panel being open, this patch only affects behavior when the panel is not displayed. It preserves the behavior of showing menu shortcuts when held down when a panel is open.

Is the intention not to have the Menu key do different things with the panel displayed and not displayed on the same device, just do different things on different types of devices?

This may be better discussed on the android-framework mailing list, though.

Patch-set: 1
Reviewer: Kenny Root <1000748@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0