Update patch set 1
Long press to show IME is only intended as a stop-gap compatibility issue for existing applications that have not been updated to be IME-aware.  It is a really crummy UI (not discoverable at all), and not something that should be relied upon.

If there is some app that doesn't allow you to show a soft keyboard, it should be updated to have an appropriate decent UI.

Patch-set: 1