Update patch set 2
Patch Set 2: No score; No score

Reading the other thread, it seems like this is meant to support a switch rather than a key event. If so, I think this is the wrong way to go about it.

For example, how will this handle the case where the user has set the switch to silent while the device is turned off? The AudioService will restore the last known settings. It seems like this needs a different treatment in the application framework, somewhat similar to the way docking works.

Patch-set: 2
Reviewer: Dave Sparks <1000387@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0