Fix a race condition determining whether password fallback mode is allowed.
The fix is in LockPatternKeyguardView, whose constructor was
firing off an asynchronous request to the AccountManager to
find out about the specifics of the account on the device.
(If it's SAML, we don't have the password in cleartext and
therefore can't use it to unlock.) Unfortunately, if the
AccountManager responds too quickly, we get the answer (in
LPKV.run()) before the UnlockScreen has even been
instantiated (later in LPKV's ctor).

The fix is to create the unlock screen first and *then* ping
the AccountManager for details.

Bug: http://b/2216308
Change-Id: Iedc84675c0ab8a001d062d806e2bee7ed1a29758