Issue 4625: Lock screen silent mode selection does not enable vibration.
This patch set fixes the issue that the silent mode selection on
the lock screen does not consider the vibration settings. This patch
allows setting the phone on vibrate mode if the vibration settings
is enabled and the user sets the phone on silent on the lock screen.
The same behavior is also fixed on the power off dialog silent mode
option.

Change-Id: I05da86e8a5e96e0bb2fb1bf59a5689e1fe6484fb
Signed-off-by: David Sobreira Marques <dpsmarques@gmail.com>