Fix 2575838: Update emergency button in AccountUnlockScreen DO NOT MERGE
AccountUnlockScreen wasn't registered to receive updates from
UpdateMonitor.  It now correctly updates the Emergency call button
to allow the user to return to the call in progress (InCallScreen).

Tested on WVGA device with keyboard:
- hitting Home during call, observing state, and pressing button to return to call.
- hitting Home while not in call to ensure it still invokes EmergencyDialer.
- above in various keyboard states.

Marked DO NOT MERGE since this will require manual merge to master due to framework restructuring.

Change-Id: Idf6bb25438336616e6453e7afe0e79f227af98d3