Update patch set 2
Patch Set 2:

Merges and builds fine. Emulator starts, and long-presses on the PC keyboard, emulated keyboard and soft keyboard have the desired behavior.

Note to reviewers: because of an internal code re-organization, this code will require some manual work to merge, and because of that this project is restricted (and nobody can +2 it by default). Please leave notes in your comments if the change is approved, so I can do the rest of the manual work around it.

Patch-set: 2