Telephony : Dial emergency phone numbers when screen is locked
The current implementation do not allow dialing emergency phone
numbers with the qwerty-keyboard when the screen is locked.

This contribution will automatically start the emergency dialer
if a valid digit (i.e. 0-9,*,+,#) is entered from the qwerty
keypad when the screen is locked. The emergency dialer appears
pre-filled with the first entered digit.

Change-Id: I136b5434f99bfab6f1908d3698e2700f7e59d036