Lockscreen: Revert patch to skip draws
This patch reverts:

"Fix for 2292713: Remove workaround that hides SlidingTab
widget while internal state is inconsistent."

The problem with this fix is that sometimes,
mIsPortrait is assigned a incorrect value,
during bootup.

If this happens, the condition to skip the
lockscreen draw is always met, and we dont see
the lockscreen on bootup.

Change-Id: I504a90549b7ff87a424ef7d6b7ce621498eeab52
Signed-off-by: Axel Haslam <axelhaslam@ti.com>