Update patch set 2
Patch Set 2:

This project is obsolete I believe.  It hasn't been tagged since Froyo, 2.2.3:

https://android.googlesource.com/platform/frameworks/policies/base/+refs

I think this project is deprecated - we don't get it with our default AOSP manifests.  The history is still there and the project still exists but the live copy of this file is now in project/base I believe.

Can you give some context?  How did you fetch this and why?  Are you using old versions of the manifest which lists this project and did you do that on purpose, etc.

Patch-set: 2
Reviewer: Robert Greenwalt <1002609@85c56323-6fa9-3386-8a01-6480fb634889>